+++ cli/command.go
@@ -3,13 +3,9 @@ package cli
 import (
 	"context"
 	"fmt"
-	"io"
 	"log"
-	"os"
 
 	"github.com/kvesta/vesta/config"
-	"github.com/kvesta/vesta/internal"
-	"github.com/kvesta/vesta/pkg/inspector"
 	"github.com/kvesta/vesta/pkg/vulnlib"
 	"github.com/spf13/cobra"
 )
@@ -26,20 +22,12 @@ var (
 	nameSpace  string
 	kubeconfig string
 	outfile    string
-	upgradeall bool
+	updateall  bool
 	skipUpdate bool
+	inside     bool
 )
 
 func Execute() error {
-	scanCmd := &cobra.Command{
-		Use:   "scan [OPTIONS]",
-		Short: "Container scan",
-	}
-
-	analyzeCmd := &cobra.Command{
-		Use:   "analyze",
-		Short: "Kubernetes analyze",
-	}
 
 	versionCmd := &cobra.Command{
 		Use:   "version",
@@ -52,12 +40,12 @@ func Execute() error {
 
 	// Upgrade vulnerability database
 	dataupgradeCmd := &cobra.Command{
-		Use:   "upgrade",
-		Short: "Upgrade vulnerability database",
+		Use:   "update",
+		Short: "Update vulnerability database",
 		Args:  NoArgs,
 		Run: func(cmd *cobra.Command, args []string) {
 			ctx := config.Ctx
-			ctx = context.WithValue(ctx, "reset", upgradeall)
+			ctx = context.WithValue(ctx, "reset", updateall)
 
 			err := vulnlib.Fetch(ctx)
 			if err != nil {
@@ -68,130 +56,13 @@ func Execute() error {
 		},
 	}
 
-	dockerAnalyze := &cobra.Command{
-		Use:   "docker",
-		Short: "analyze docker container",
-		Run: func(cmd *cobra.Command, args []string) {
-			ctx := config.Ctx
-			ctx = context.WithValue(ctx, "output", outfile)
+	dataupgradeCmd.Flags().BoolVarP(&updateall, "all", "a", false, "Reset the database")
 
-			internal.DoInspectInDocker(ctx)
-		},
-	}
-
-	kubernetesAnalyze := &cobra.Command{
-		Use:   "k8s",
-		Short: "analyze configure of kubernetes",
-		Run: func(cmd *cobra.Command, args []string) {
-			ctx := config.Ctx
-			ctx = context.WithValue(ctx, "nameSpace", nameSpace)
-			ctx = context.WithValue(ctx, "kubeconfig", kubeconfig)
-			ctx = context.WithValue(ctx, "output", outfile)
-
-			internal.DoInspectInKubernetes(ctx)
-		},
-	}
-
-	imageCheck := &cobra.Command{
-		Use:   "image",
-		Short: "input from image",
-		Run: func(cmd *cobra.Command, args []string) {
-
-			if len(args) < 1 {
-				fmt.Println("Require at least 1 argument.")
-				os.Exit(1)
-			}
-
-			ctx := config.Ctx
-			ctx = context.WithValue(ctx, "tarType", "image")
-			ctx = context.WithValue(ctx, "output", outfile)
-			ctx = context.WithValue(ctx, "skip", skipUpdate)
-
-			var tarIO []io.ReadCloser
-
-			if tarFile == "" {
-				var err error
-				tarIO, err = inspector.GetTarFromID(ctx, args[0])
-
-				if err != nil {
-					os.Exit(1)
-				}
-
-			}
-
-			if tarFile == "" && len(tarIO) < 1 {
-				log.Printf("Can not get tarfile. " +
-					"Make sure that you have the right image ID " +
-					"or use -f to get from tar file")
-				return
-			}
-			internal.DoScan(ctx, tarFile, tarIO)
-		},
-	}
-
-	containerCheck := &cobra.Command{
-		Use:   "container",
-		Short: "input from inspector",
-		Run: func(cmd *cobra.Command, args []string) {
-
-			if len(args) < 1 {
-				fmt.Println("Require at least 1 argument.")
-				os.Exit(1)
-			}
-
-			ctx := config.Ctx
-			ctx = context.WithValue(ctx, "tarType", "container")
-			ctx = context.WithValue(ctx, "output", outfile)
-			ctx = context.WithValue(ctx, "skip", skipUpdate)
-
-			var tarIO []io.ReadCloser
-
-			if tarFile == "" {
-				var err error
-				tarIO, err = inspector.GetTarFromID(ctx, args[0])
-
-				if err != nil {
-					os.Exit(1)
-				}
-
-			}
-
-			if tarFile == "" && len(tarIO) < 1 {
-				log.Printf("Can not get tarfile. " +
-					"Make sure that you have the right container ID" +
-					"or use -f to get from tar file")
-				return
-			}
-
-			internal.DoScan(ctx, tarFile, tarIO)
-		},
-	}
-
-	imageCheck.Flags().StringVarP(&tarFile, "file", "f", "", "path of tar file")
-	imageCheck.Flags().StringVarP(&outfile, "output", "o", "output", "output file location")
-	imageCheck.Flags().BoolVar(&skipUpdate, "skip", false, "skip the updating")
-
-	containerCheck.Flags().StringVarP(&tarFile, "file", "f", "", "path of tar file")
-	containerCheck.Flags().StringVarP(&outfile, "output", "o", "output", "output file location")
-	containerCheck.Flags().BoolVar(&skipUpdate, "skip", false, "skip the updating")
-
-	kubernetesAnalyze.Flags().StringVarP(&nameSpace, "ns", "n", "all", "specific namespace")
-	kubernetesAnalyze.Flags().StringVar(&kubeconfig, "kubeconfig", "default", "specific configure file")
-	kubernetesAnalyze.Flags().StringVarP(&outfile, "output", "o", "output", "output file location")
-
-	dockerAnalyze.Flags().StringVarP(&outfile, "output", "o", "output", "output file location")
-
-	dataupgradeCmd.Flags().BoolVarP(&upgradeall, "all", "a", false, "Reset the database")
-
-	scanCmd.AddCommand(imageCheck)
-	scanCmd.AddCommand(containerCheck)
-
-	analyzeCmd.AddCommand(dockerAnalyze)
-	analyzeCmd.AddCommand(kubernetesAnalyze)
-
-	rootCmd.AddCommand(scanCmd)
-	rootCmd.AddCommand(analyzeCmd)
 	rootCmd.AddCommand(dataupgradeCmd)
 	rootCmd.AddCommand(versionCmd)
+
+	analyze()
+	scan()
+
 	return rootCmd.Execute()
 }